+++ spec/spec_helper.rb
@@ -2,7 +2,7 @@ require 'rspec'
 require 'time_difference'
 
 RSpec.configure do |config|
-  config.formatter     = 'documentation'
+  config.formatter = 'documentation'
 
   # Configure Timezone for proper tests
   ENV['TZ'] = 'UTC'